reasoning-engine/pyproject.toml
[project]
name = "reasoning-engine"
version = "0.1.0"
description = "Reasoning engine with ReAct prompting for AetherOS voice agent"
requires-python = ">=3.10"
dependencies = [
    "llama-cpp-python>=0.2.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=24.1.0",
    "ruff>=0.1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
